requirements.txt
requests 
pandas 
psycopg2 
sqlalchemy
python-dotenv